Extract screenshot taking logic into DefaultScreenshotStrategy
