+++ src/main/java/com/randomscreenshot/RandomScreenshotPlugin.java
@@ -1,14 +1,8 @@
 package com.randomscreenshot;
 
 import com.google.inject.Provides;
-import java.util.Random;
 import javax.inject.Inject;
 import lombok.extern.slf4j.Slf4j;
-import net.runelite.api.Client;
-import net.runelite.api.GameState;
-import net.runelite.api.events.GameTick;
-import net.runelite.api.widgets.ComponentID;
-import net.runelite.api.widgets.Widget;
 import net.runelite.client.config.ConfigManager;
 import net.runelite.client.eventbus.Subscribe;
 import net.runelite.client.plugins.Plugin;
@@ -25,16 +19,11 @@ public class RandomScreenshotPlugin extends Plugin
 	static final String CONFIG_GROUP_KEY = "randomscreenshot";
 
 	@Inject
-	private RandomScreenshotConfig config;
-
-	@Inject
 	private ScreenshotUtil screenShotUtil;
 
+	// TODO: Instead of injection, this could be chosen dynamically via config.
 	@Inject
-	private Client client;
-
-	@Inject
-	private RandomUtil rand;
+	private ScreenshotStrategy screenshotStrategy;
 
 	@Provides
 	RandomScreenshotConfig getConfig(ConfigManager configManager)
@@ -43,38 +32,15 @@ public class RandomScreenshotPlugin extends Plugin
 	}
 
 	@Subscribe
-	public void onGameTick(GameTick event)
+	public void onGameTick()
 	{
-		if (shouldTakeScreenshot())
+		if (screenshotStrategy.shouldTakeScreenshot())
 		{
 			takeScreenshot();
 		}
 	}
 
-	/* TODO: Create a decider interface so that decision strategy can be made modular. */
-	boolean shouldTakeScreenshot() {
-		if (isBankPinContainerVisible() || isOnLoginScreen()) {
-			return false;
-		}
-
-		return rand.randInt(config.sampleWeight()) == 0;
-	}
-
 	private void takeScreenshot() {
 		screenShotUtil.takeScreenshot();
-	}
-
-	private boolean isBankPinContainerVisible()
-	{
-		Widget pinContainer = client.getWidget(ComponentID.BANK_PIN_CONTAINER);
-		if (pinContainer == null) {
-			return false;
-		}
-
-		return !pinContainer.isSelfHidden();
-	}
-
-	private boolean isOnLoginScreen() {
-		return client.getGameState() == GameState.LOGIN_SCREEN;
 	}
 }
