+++ src/main/java/com/randomscreenshot/ScreenshotStrategy.java
@@ -1,0 +1,6 @@
+package com.randomscreenshot;
+
+public interface ScreenshotStrategy
+{
+	boolean shouldTakeScreenshot();
+}
